hit1/centralizador/requirements.txt
pika==1.3.2
Pillow==10.2.0
numpy==1.26.4